+++ src/LiveComponent/src/Metadata/LiveComponentMetadata.php
@@ -22,12 +22,12 @@ class LiveComponentMetadata
 {
     public function __construct(
         private ComponentMetadata $componentMetadata,
-        /** @var list<LivePropMetadata|LegacyLivePropMetadata> */
+        /** @var list<LivePropMetadata> */
         private array $livePropsMetadata,
     ) {
         uasort(
             $this->livePropsMetadata,
-            static fn (LivePropMetadata|LegacyLivePropMetadata $a, LivePropMetadata|LegacyLivePropMetadata $b) => $a->hasModifier() <=> $b->hasModifier()
+            static fn (LivePropMetadata $a, LivePropMetadata $b) => $a->hasModifier() <=> $b->hasModifier()
         );
     }
 
@@ -37,7 +37,7 @@ class LiveComponentMetadata
     }
 
     /**
-     * @return list<LivePropMetadata|LegacyLivePropMetadata>
+     * @return list<LivePropMetadata>
      */
     public function getAllLivePropsMetadata(?object $component): iterable
     {
@@ -55,7 +55,7 @@ class LiveComponentMetadata
      */
     public function getOnlyPropsThatAcceptUpdatesFromParent(array $inputProps): array
     {
-        $writableProps = array_filter($this->livePropsMetadata, static function (LivePropMetadata|LegacyLivePropMetadata $livePropMetadata) {
+        $writableProps = array_filter($this->livePropsMetadata, static function (LivePropMetadata $livePropMetadata) {
             return $livePropMetadata->acceptUpdatesFromParent();
         });